Avoid possible TypeError Exception
Exception: unlink(): Argument #1 ($filename) must be of type string, bool given
In [/home/wotanabx/webseiten/wow-das/vendor/cakephp/cakephp/src/Cache/Engine/FileEngine.php, line 227]

2022-03-11 02:32:30 error: [TypeError] unlink(): Argument #1 ($filename) must be of type string, bool given in /home/wotanabx/webseiten/wow-das/vendor/cakephp/cakephp/src/Cache/Engine/FileEngine.php on line 227

+++ src/Cache/Engine/FileEngine.php
@@ -222,6 +222,10 @@ class FileEngine extends CacheEngine
         /** @psalm-suppress PossiblyNullReference */
         $path = $this->_File->getRealPath();
         $this->_File = null;
+        
+        if ($path === false) {
+            return false;
+        }
 
         // phpcs:disable
         return @unlink($path);
